hw6
