--- src/home_work_6/ComparatorList.java
@@ -1,11 +1,0 @@
-package home_work_6;
-
-import java.util.*;
-
-public class ComparatorList implements Comparator <Map.Entry<String,Integer>> {
-    @Override
-    public int compare(Map.Entry<String,Integer> o1, Map.Entry<String,Integer> o2) {
-        return o2.getValue()-o1.getValue();
-
-    }
-}
