--- src/home_work_6/ISearchEngine.java
@@ -1,8 +1,0 @@
-
-package home_work_6;
-
-public interface ISearchEngine {
-  long search(String text,String word);
-
-
-}
