+++ src/home_work_6/api/ISearchEngine.java
@@ -1,0 +1,14 @@
+
+package home_work_6;
+
+
+public interface ISearchEngine {
+  /**
+   *метод нахождения колличества подобных слов в тексте
+   * @param text текст в котором осуществляем поиск
+   * @param word искомое слово
+   */
+  long search(String text,String word);
+
+
+}
