+++ src/home_work_6/runners/BookGenerator.java
@@ -1,5 +1,6 @@
 //7.Написать новый класс с точкой входа. При запуске программы пользователь должен передать адрес папки,
-// из которой мы будет считывать текстовые файлы. Наполнить данную директорию книгами, 100 книг мнинимум
+// из которой мы будет считывать текстовые файлы. Наполнить данную директорию книгами, 100 книг мнинимум<---
+//-<<<сделал только до сюда
 package home_work_6;
 import java.io.BufferedWriter;
 import java.io.File;
