+++ src/home_work_6/runners/Main1.java
@@ -13,10 +13,10 @@ import java.util.regex.Pattern;
 
 public class Main1 {
     public static void main(String[] args) {
-              Set<String> set = new HashSet<String>();
+        Set<String> set = new HashSet<String>();
         try {
-            BufferedReader text = new BufferedReader(new FileReader("src/Война и мир.txt")); // чтение текста с файла
-               String st = "";
+            BufferedReader text = new BufferedReader(new FileReader("C:\\Users\\intel i7\\IdeaProjects\\JD1\\HomeWork\\src\\source\\Война и мир.txt")); // чтение текста с файла
+            String st = "";
             while ((st = text.readLine()) != null) { //проверяем построчно есть ли заполненная строка теста и сохраняем эту строку в st
                 StringTokenizer tokenizer = new StringTokenizer(st, " .!?(){}[]<>#*,:;-\'\"/"); //раскладывает строку на части - слова по делиметру
                 while (tokenizer.hasMoreTokens()) {  // функция наличия слов-лексем
